server side better management, routes and other in outer files
