+++ book-visit-app-server/.env
@@ -1,3 +1,6 @@
+NODE_SERVER_IP=http://localhost:5000
+NODE_SERVER_PORT=5000
+
 NODE_GMAIL_SENDER_MAIL=dawid.book.app@gmail.com
 NODE_GMAIL_SENDER_PASSWORD=zaq1@WSXQwEr!
