--- book-visit-app-server/env
@@ -1,12 +1,0 @@
-NODE_GMAIL_SENDER_MAIL=dawid.book.app@gmail.com
-NODE_GMAIL_SENDER_PASSWORD=zaq1@WSXQwEr!
-
-REACT_APP_FIREBASE_API_KEY=AIzaSyDzG340nqg5aAfp_9-7g6R3MZajtPXCvRU
-REACT_APP_FIREBASE_AUTH_DOMAIN=book-visit-app.firebaseapp.com
-REACT_APP_FIREBASE_DATABEASE_URL=https://book-visit-app-default-rtdb.europe-west1.firebasedatabase.app/
-REACT_APP_FIREBASE_PROJECT_ID=book-visit-app
-REACT_APP_FIREBASE_STORAGE_BUCKET=book-visit-app.appspot.com
-REACT_APP_FIREBASE_MESSAGING_SENDER_ID=566914259425
-REACT_APP_FIREBASE_APP_ID=1:566914259425:web:57b5084311c2f89166daa0
-REACT_APP_FIREBASE_MEASUREMENT_ID=G-ZRGZYSXZG4
-
